pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=3.1.2",
    "django>=6.0.3",
    "django-bootstrap5>=26.2",
    "gunicorn>=25.3.0",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.2",
    "whitenoise>=6.12.0",
    "django-filter>=23.5",
    "rollbar>=0.16.0",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
    "django-bootstrap4>=26.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["task_manager"]

[dependency-groups]
dev = [
    "flake8>=7.3.0",
]
